+++ tests/system/fixtures/not_to_file.sh
@@ -1,0 +1,3 @@
+echo "Notification received"
+
+echo "Notification received" > received.txt
