+++ tests/unit/fixtures/good_listeners/command_json_listener.yaml
@@ -7,5 +7,5 @@ listeners:
       number: AZ203
     triggers:
       - type: command
-        working_dir: tests/integration/fixtures
+        working_dir: tests/unit/fixtures
         command: ./my_script.sh --json ${json}
