+++ tests/unit/fixtures/listener_schema.json
@@ -1,0 +1,1 @@
+{"version": 0.1, "flight": {"endpoint": [{"engine": ["etcd_rest", "etcd_grpc", "file_based"], "base": "/tmp/aviso/flight/{country}/", "stem": "{date}/{airport}/{number}", "admin": "/tmp/admin/{country}"}], "request": {"date": [{"canonic": "%Y%m%d", "type": "DateHandler"}], "country": [{"canonic": "lower", "type": "StringHandler"}], "airport": [{"canonic": "upper", "type": "StringHandler"}], "number": [{"type": "StringHandler"}]}}}
